requirements.txt
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
beautifulsoup4>=4.12.0
requests>=2.31.0
selenium>=4.15.0
markdown>=3.5.0
weasyprint>=60.0
pdfkit>=1.0.0
tinydb>=4.8.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
chainlit>=0.7.0
arxiv>=2.1.0
scholarly>=1.7.0
newsapi-python>=0.2.6
semantic-kernel>=0.4.0
pydantic>=2.5.0
aiohttp>=3.9.0 